+++ src/main/java/com/home/userbookwishlist/service/UserService.java
@@ -1,22 +1,25 @@
 package com.home.userbookwishlist.service;
 
+import com.google.inject.Inject;
 import com.home.userbookwishlist.dao.UserDao;
+import com.home.userbookwishlist.dao.UserWishListDao;
 import com.home.userbookwishlist.externalobject.ExternalUser;
-import com.home.userbookwishlist.sqlitedb.SqliteConnection;
 import com.home.userbookwishlist.model.User;
-import com.home.userbookwishlist.userapi.RemoveUser;
-import com.home.userbookwishlist.sqlitedb.SQLConnectionObject;
+import com.home.userbookwishlist.model.UserWishList;
+import org.springframework.util.CollectionUtils;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import java.util.List;
 
-@Path("/userService")
 public class UserService {
 
-
+    private final UserWishListDao userWishListDao;
     private final UserDao userDao;
 
-    public UserService(final UserDao userDao) {
+    @Inject
+    public UserService(final UserWishListDao userWishListDao, final UserDao userDao) {
+        this.userWishListDao = userWishListDao;
         this.userDao = userDao;
     }
 
@@ -25,8 +28,18 @@ public class UserService {
         userDao.addUser(user);
     }
 
-    public void removeUser(final int userId) {
+    public String removeUser(final int userId) {
+        if(isUserNotExist(userId)) {
+            return String.format("user_id %d doesn't exist in our system.", userId);
+        }
+
+        List<UserWishList> userWishListResult = userWishListDao.getUserWishList(userId);
+        if(!CollectionUtils.isEmpty(userWishListResult)) {
+            userWishListResult.stream()
+                    .forEach(userWishList -> userWishListDao.removeBookFromWishList(userWishList));
+        }
         userDao.removeUser(userId);
+        return String.format("user_id %d has been removed.", userId);
     }
 
     private User convertToUser(final ExternalUser externalUser) {
@@ -34,50 +47,11 @@ public class UserService {
                 .email(externalUser.getEmail())
                 .firstName(externalUser.getFirstName())
                 .lastName(externalUser.getLastName())
-                .password(externalUser.getPassword())
                 .userId(externalUser.getUserId())
                 .build();
     }
 
-
-//    @GET
-//    @Produces(MediaType.APPLICATION_JSON)
-//    @Path("/getUser")
-//    public Book read(@QueryParam("user_mail") String user_email) {
-//        GetBookInWishListActivity viewBookInWishList = new GetBookInWishListActivity();
-//        SqliteConnection.connect();
-//        return  viewBookInWishList.retrieve(user_email);
-////        String bookDetails= viewBookInWishList.retrieve(isbn);
-////            return "Book details for the isbn number: " +
-////                    bookDetails;
-//
-//    }
-
-//    @PUT
-//    @Path("/updateUser")
-//    @Consumes(MediaType.APPLICATION_XML)
-//    @Produces(MediaType.TEXT_PLAIN)
-////fetch the user name from user activity
-//    public String updateBook(@ModelAttribute Book book) {
-//        UpdateBookInWishListActivity updateBookInWishList = new UpdateBookInWishListActivity();
-//        if(SQLConnectionObject.conn == null){
-//            SqliteConnection.connect();
-//        }
-//        updateBookInWishList.update(book.getIsbn(), book.getAuthor() ,
-//                book.getTitle() , book.getDateOfPublication(), book.getUser_email());
-//        return "User updated";
-//    }
-
-    @GET
-    @Path("/deleteUser")
-    @Produces(MediaType.TEXT_PLAIN)
-    public String delete(@QueryParam("userId") final String userId ) {
-
-        /*final RemoveUser removeUser = new RemoveUser();
-        if(SQLConnectionObject.conn == null){
-            SqliteConnection.connect();
-        }
-        removeUser.remove(userId);*/
-        return "User and corresponding book list has been deleted";
+    private boolean isUserNotExist(final int userId) {
+        return userDao.getUser(userId).size() == 0;
     }
 }
